Make the feature tests pass
