+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/transformer/Server.scala
@@ -2,18 +2,11 @@ package uk.ac.wellcome.platform.transformer
 
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finatra.http.HttpServer
-import com.twitter.finatra.http.filters.{
-  CommonFilters,
-  LoggingMDCFilter,
-  TraceIdMDCFilter
-}
+import com.twitter.finatra.http.filters.{CommonFilters, LoggingMDCFilter, TraceIdMDCFilter}
 import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.finatra.controllers.ManagementController
-import uk.ac.wellcome.transformer.modules.{
-  TransformableParserModule,
-  TransformerWorkerModule
-}
+import uk.ac.wellcome.transformer.modules.{TransformerModule, TransformerWorkerModule}
 
 object ServerMain extends Server
 
@@ -28,8 +21,8 @@ class Server extends HttpServer {
     SQSConfigModule,
     SNSConfigModule,
     SNSClientModule,
-    TransformableParserModule,
-    TransformerWorkerModule
+    TransformerWorkerModule,
+    TransformerModule
   )
   override def configureHttp(router: HttpRouter) {
     router
